move export building logic out of controller and into new PublicWaterSystemExporter
